+++ packages/stack-sc/src/index.client.ts
@@ -1,1 +1,1 @@
-export { cookies } from 'next/headers';
+export const cookies = undefined;