+++ packages/stack-sc/src/index.combined.ts
@@ -1,0 +1,4 @@
+import * as server from "./index.server";
+import * as client from "./index.client";
+
+export declare const cookies: typeof server.cookies | typeof client.cookies;