+++ packages/stack-server/src/middleware.tsx
@@ -10,6 +10,7 @@ const corsAllowedRequestHeaders = [
   'content-type',
   'x-stack-project-id',
   'x-stack-override-error-status',
+  'x-stack-random-nonce',  // used to forcefully disable some caches
   
   // Project auth
   'x-stack-request-type',