+++ packages/stack-shared/src/interface/clientInterface.ts
@@ -7,6 +7,8 @@ import { AsyncStore, ReadonlyAsyncStore } from '../utils/stores';
 import { KnownError, KnownErrors } from '../known-errors';
 import { StackAssertionError } from '../utils/errors';
 import { ProjectUpdateOptions } from './adminInterface';
+import { cookies } from '@stackframe/stack-sc';
+import { generateSecureRandomString } from '../utils/crypto';
 
 type UserCustomizableJson = {
   readonly displayName: string | null,
@@ -320,6 +322,9 @@ export class StackClientInterface {
       tokenObj = await tokenStore.getOrWait();
     }
 
+    // all requests should be dynamic to prevent Next.js caching
+    cookies?.();
+
     const url = this.getApiUrl() + path;
     const params: RequestInit = {
       /**
@@ -344,8 +349,10 @@ export class StackClientInterface {
         ...'projectOwnerTokens' in this.options ? {
           "X-Stack-Admin-Access-Token": (await this.options.projectOwnerTokens?.getOrWait())?.accessToken ?? "",
         } : {},
+        "X-Stack-Random-Nonce": generateSecureRandomString(),
         ...options.headers,
       },
+      cache: "no-store",
     };
 
     const rawRes = await fetch(url, params);