+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,11 @@
 # e2e-tests
 
+## 0.1.1
+
+### Patch Changes
+
+- @stackframe package namespace
+
 ## 0.1.0
 
 ### Minor Changes