+++ apps/partial-prerendering/CHANGELOG.md
@@ -1,5 +1,20 @@
 # partial-prerendering
 
+## 0.2.2
+
+### Patch Changes
+
+- @stackframe package namespace
+- Updated dependencies
+  - @stackframe/stack@2.4.2
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.4.1
+
 ## 0.2.0
 
 ### Minor Changes