+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack-shared
 
+## 2.4.1
+
+### Patch Changes
+
+- Teams, permissions and RBAC
+- Updated dependencies
+  - @stackframe/stack-sc@1.5.1
+
 ## 2.4.0
 
 ### Minor Changes